Cleaning up Parser API
Renamed `cp<X>WS` stuff to `p<X>`, and when there was a counterpart
`cp<X>` (no `WS` suffix), these got renamed to `p<X>_DEP` because they
will be deprecated in the future.

In the future, everything will use `Parser t` when the token type is
generic, and `Parser TokenWSBasic` when the token type is specific to
the default language parsing infrastructure.